git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@2167 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/sql/wbcommands/TableCopy.java
@@ -110,6 +110,11 @@ class TableCopy
 			{
 				targetId = new TableIdentifier(targettable);
 			}
+
+			if (targetId == null)
+			{
+				throw new TableNotFoundException(targettable);
+			}
 			copier.copyFromTable(sourceConnection, targetConnection, srcTable, targetId, mapping, where, createTableType, dropTable, ignoreDropError, skipTargetCheck);
 		}
 		else
@@ -133,6 +138,10 @@ class TableCopy
 					queryCols.get(i).setColumnName(cols.get(i).getColumnName());
 					queryCols.get(i).setColumnAlias(null);
 				}
+			}
+			if (targetId == null)
+			{
+				targetId = new TableIdentifier(targettable);
 			}
 			copier.copyFromQuery(sourceConnection, targetConnection, sourcequery, targetId, queryCols, createTableType, dropTable, ignoreDropError, skipTargetCheck);
 		}
